Fixes for UI and PDF Library
Fixed UI Layout
Changed PDF library to GhostScript, still experiencing issues
