+++ ARTAP-Client/PDFToImage/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Ghostscript.NET" version="1.2.1" targetFramework="net452" />
+</packages>
