+++ ARTAP-Client/PDFViewer/PDFViewerDialog.xaml
@@ -1,0 +1,25 @@
+﻿<Window x:Class="PDFViewer.PDFViewerDialog"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:PDFViewer"
+        mc:Ignorable="d"
+        Title="PDFViewerDialog" Height="700" Width="500">
+    <Grid HorizontalAlignment="Stretch" Margin="0,0,0,0" VerticalAlignment="Stretch">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="Auto"/>
+            <ColumnDefinition Width="*"/>
+        </Grid.ColumnDefinitions>
+        <Image x:Name="imagePageDisplay" Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="3" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Height="Auto" Width="Auto"/>
+        <Button x:Name="buttonBack" Grid.Row="1" Grid.Column="0" Height="30" Width="40" VerticalAlignment="Center" HorizontalAlignment="Right" Content="&lt;" Click="onClickButtonBack"/>
+        <TextBox x:Name="textBoxPageNumber" Grid.Row="1" Grid.Column="1" Height="30" Width="50" VerticalAlignment="Center" HorizontalAlignment="Center" Margin="10" MaxLines="1" TextChanged="onTextChangedTextBoxPageNumber"/>
+        <Button x:Name="buttonForward" Grid.Row="1" Grid.Column="2" Height="30" Width="40" VerticalAlignment="Center" HorizontalAlignment="Left" Content="&gt;" Click="onClickButtonForward"/>
+        <Button x:Name="buttonSelect" Grid.Row="1" Grid.Column="2" Height="30" Width="50" VerticalAlignment="Center" HorizontalAlignment="Right" Margin="10" Content="Select" Click="onClickButtonSelect"/>
+    </Grid>
+</Window>
