--- PDFTool/PDFTool.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PDFToImage", "ClassLibrary1\PDFToImage.csproj", "{0C7E4FB9-50D4-4E46-BA97-7B6CFF41351F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PDFViewer", "PDFViewer\PDFViewer.csproj", "{378340EF-F1F5-490E-8D0C-FBF22B811F25}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0C7E4FB9-50D4-4E46-BA97-7B6CFF41351F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0C7E4FB9-50D4-4E46-BA97-7B6CFF41351F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C7E4FB9-50D4-4E46-BA97-7B6CFF41351F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0C7E4FB9-50D4-4E46-BA97-7B6CFF41351F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{378340EF-F1F5-490E-8D0C-FBF22B811F25}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{378340EF-F1F5-490E-8D0C-FBF22B811F25}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{378340EF-F1F5-490E-8D0C-FBF22B811F25}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{378340EF-F1F5-490E-8D0C-FBF22B811F25}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
